+++ packages/atomic/src/components.d.ts
@@ -5,11 +5,11 @@
  * It contains typing information for all components that exist in this project.
  */
 import { HTMLStencilElement, JSXBase } from "@stencil/core/internal";
-import { CategoryFacet, CategoryFacetState, Engine, Facet, FacetState, LogLevel, Result, ResultTemplateCondition } from "@coveo/headless";
-import { Bindings, I18nState } from "./utils/initialization-utils";
+import { Engine, LogLevel, Result, ResultTemplateCondition } from "@coveo/headless";
+import { Bindings } from "./utils/initialization-utils";
 import { i18n } from "i18next";
 import { InitializationOptions } from "./components/atomic-search-interface/atomic-search-interface";
-import { Suggestion } from "./components/searchbox/base-search";
+import { BaseFacetSearchResult } from "@coveo/headless/dist/api/search/facet-search/base/base-facet-search-response";
 export namespace Components {
     interface AtomicBreadcrumbManager {
         "categoryDivider": string;
@@ -114,17 +114,14 @@ export namespace Components {
         "_id": string;
         /**
           * Whether the submit button should be placed before the input
-          * @default false
          */
         "leadingSubmitButton": boolean;
         /**
           * Maximum number of suggestions to display
-          * @default 5
          */
         "numberOfSuggestions": number;
         /**
           * The placeholder for the search box input
-          * @default ''
          */
         "placeholder": string;
     }
@@ -169,19 +166,9 @@ export namespace Components {
         "hasActiveValues": boolean;
         "label": string;
     }
-    interface BaseSearch {
-        "hideSubmit": boolean;
-        "id": string;
-        "leadingSubmitButton": boolean;
-        "moreValuesAvailable": boolean;
-        "placeholder": string;
-        "strings": I18nState;
-        "suggestionValues": Suggestion[];
-    }
     interface FacetSearch {
-        "_id": string;
-        "facet": Facet | CategoryFacet;
-        "facetState": FacetState | CategoryFacetState;
+        "facetSearchResults": BaseFacetSearchResult[];
+        "moreValuesAvailable": boolean;
     }
     interface FacetValue {
         "isSelected": boolean;
@@ -358,12 +345,6 @@ declare global {
         prototype: HTMLBaseFacetElement;
         new (): HTMLBaseFacetElement;
     };
-    interface HTMLBaseSearchElement extends Components.BaseSearch, HTMLStencilElement {
-    }
-    var HTMLBaseSearchElement: {
-        prototype: HTMLBaseSearchElement;
-        new (): HTMLBaseSearchElement;
-    };
     interface HTMLFacetSearchElement extends Components.FacetSearch, HTMLStencilElement {
     }
     var HTMLFacetSearchElement: {
@@ -405,7 +386,6 @@ declare global {
         "atomic-tab": HTMLAtomicTabElement;
         "atomic-text": HTMLAtomicTextElement;
         "base-facet": HTMLBaseFacetElement;
-        "base-search": HTMLBaseSearchElement;
         "facet-search": HTMLFacetSearchElement;
         "facet-value": HTMLFacetValueElement;
     }
@@ -511,17 +491,14 @@ declare namespace LocalJSX {
         "_id"?: string;
         /**
           * Whether the submit button should be placed before the input
-          * @default false
          */
         "leadingSubmitButton"?: boolean;
         /**
           * Maximum number of suggestions to display
-          * @default 5
          */
         "numberOfSuggestions"?: number;
         /**
           * The placeholder for the search box input
-          * @default ''
          */
         "placeholder"?: string;
     }
@@ -565,26 +542,12 @@ declare namespace LocalJSX {
         "label": string;
         "onDeselectAll"?: (event: CustomEvent<void>) => void;
     }
-    interface BaseSearch {
-        "hideSubmit"?: boolean;
-        "id": string;
-        "leadingSubmitButton"?: boolean;
-        "moreValuesAvailable"?: boolean;
-        "onClear"?: (event: CustomEvent<void>) => void;
-        "onHideSuggestions"?: (event: CustomEvent<void>) => void;
-        "onSelectValue"?: (event: CustomEvent<number>) => void;
-        "onShowMoreResults"?: (event: CustomEvent<void>) => void;
-        "onShowSuggestions"?: (event: CustomEvent<void>) => void;
-        "onSubmit"?: (event: CustomEvent<void>) => void;
-        "onTextChange"?: (event: CustomEvent<string>) => void;
-        "placeholder"?: string;
-        "strings": I18nState;
-        "suggestionValues": Suggestion[];
-    }
     interface FacetSearch {
-        "_id"?: string;
-        "facet": Facet | CategoryFacet;
-        "facetState": FacetState | CategoryFacetState;
+        "facetSearchResults": BaseFacetSearchResult[];
+        "moreValuesAvailable": boolean;
+        "onFacetSearch"?: (event: CustomEvent<string>) => void;
+        "onResultSelected"?: (event: CustomEvent<BaseFacetSearchResult>) => void;
+        "onShowMoreResults"?: (event: CustomEvent<void>) => void;
     }
     interface FacetValue {
         "isSelected": boolean;
@@ -621,7 +584,6 @@ declare namespace LocalJSX {
         "atomic-tab": AtomicTab;
         "atomic-text": AtomicText;
         "base-facet": BaseFacet;
-        "base-search": BaseSearch;
         "facet-search": FacetSearch;
         "facet-value": FacetValue;
     }
@@ -658,7 +620,6 @@ declare module "@stencil/core" {
             "atomic-tab": LocalJSX.AtomicTab & JSXBase.HTMLAttributes<HTMLAtomicTabElement>;
             "atomic-text": LocalJSX.AtomicText & JSXBase.HTMLAttributes<HTMLAtomicTextElement>;
             "base-facet": LocalJSX.BaseFacet & JSXBase.HTMLAttributes<HTMLBaseFacetElement>;
-            "base-search": LocalJSX.BaseSearch & JSXBase.HTMLAttributes<HTMLBaseSearchElement>;
             "facet-search": LocalJSX.FacetSearch & JSXBase.HTMLAttributes<HTMLFacetSearchElement>;
             "facet-value": LocalJSX.FacetValue & JSXBase.HTMLAttributes<HTMLFacetValueElement>;
         }