+++ packages/headless/doc-parser/mocks/mock-func-entity.ts
@@ -1,0 +1,13 @@
+import {FuncEntity} from '../src/entity';
+
+export function buildMockFuncEntity(
+  config: Partial<FuncEntity> = {}
+): FuncEntity {
+  return {
+    desc: '',
+    name: '',
+    params: [],
+    returnType: '',
+    ...config,
+  };
+}